Move TransactionCommitLockInfo to impl as it is not (yet) part of the public API
